# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SHELLOC Backend API (Reloaded)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/robot-status/{robot_id}": {
      "post": {
        "tags": [
          "Robot Status"
        ],
        "summary": "Update Robot Status",
        "description": "Upserts the robot status document.\nRequires X-API-Key authentication.",
        "operationId": "update_robot_status_api_robot_status__robot_id__post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "robot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Robot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RobotStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Robot Status"
        ],
        "summary": "Get Robot Status",
        "description": "Retrieves the latest robot status. No authentication required.",
        "operationId": "get_robot_status_api_robot_status__robot_id__get",
        "parameters": [
          {
            "name": "robot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Robot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Robot Status"
        ],
        "summary": "Dispatch Robot",
        "description": "Dispatches the robot to a specific waypoint by setting target_waypoint_id.",
        "operationId": "dispatch_robot_api_robot_status__robot_id__patch",
        "parameters": [
          {
            "name": "robot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Robot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RobotStatusDispatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/waypoints/robot/{robot_id}": {
      "get": {
        "tags": [
          "Waypoints"
        ],
        "summary": "List Waypoints By Robot",
        "operationId": "list_waypoints_by_robot_api_waypoints_robot__robot_id__get",
        "parameters": [
          {
            "name": "robot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Robot Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaypointOut"
                  },
                  "title": "Response List Waypoints By Robot Api Waypoints Robot  Robot Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RobotStatusDispatch": {
        "properties": {
          "target_waypoint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Waypoint Id",
            "description": "ObjectId string of the waypoint to dispatch the robot to. Null to cancel active target."
          }
        },
        "type": "object",
        "title": "RobotStatusDispatch"
      },
      "RobotStatusOut": {
        "properties": {
          "operation_mode": {
            "type": "string",
            "enum": [
              "autonomous",
              "manual"
            ],
            "title": "Operation Mode",
            "description": "The current operation mode of the robot"
          },
          "gps_signal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "good",
                  "weak",
                  "none"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Gps Signal",
            "description": "Quality of the GPS signal"
          },
          "current_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Lat",
            "description": "Current latitude"
          },
          "current_lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Lng",
            "description": "Current longitude"
          },
          "heading_degrees": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 360.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading Degrees",
            "description": "Robot heading in degrees (0-360)"
          },
          "battery_percent": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Battery Percent",
            "description": "Battery percentage (0-100)"
          },
          "points_treated_today": {
            "type": "integer",
            "title": "Points Treated Today",
            "description": "Number of waypoints treated today",
            "default": 0
          },
          "target_waypoint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Waypoint Id",
            "description": "ObjectId string of the target waypoint, if dispatched"
          },
          "mission_state": {
            "type": "string",
            "enum": [
              "idle",
              "navigating",
              "baseline_evaluating",
              "dispensing_flocculant",
              "incubating_15m",
              "mesh_biochar_filtering",
              "post_evaluating",
              "adaptive_stabilization",
              "completed",
              "failsafe_buoyancy",
              "manual_override"
            ],
            "title": "Mission State",
            "description": "The robot's current mission phase",
            "default": "idle"
          },
          "robot_id": {
            "type": "string",
            "title": "Robot Id",
            "description": "The unique identifier for the robot"
          },
          "last_sync": {
            "type": "string",
            "format": "date-time",
            "title": "Last Sync",
            "description": "Server-computed UTC timestamp of the last heartbeat"
          },
          "overall_status": {
            "type": "string",
            "title": "Overall Status",
            "description": "Server-computed overall status (e.g., operational, low_battery, gps_lost, degraded)"
          }
        },
        "type": "object",
        "required": [
          "operation_mode",
          "battery_percent",
          "robot_id",
          "last_sync",
          "overall_status"
        ],
        "title": "RobotStatusOut"
      },
      "RobotStatusUpdate": {
        "properties": {
          "operation_mode": {
            "type": "string",
            "enum": [
              "autonomous",
              "manual"
            ],
            "title": "Operation Mode",
            "description": "The current operation mode of the robot"
          },
          "gps_signal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "good",
                  "weak",
                  "none"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Gps Signal",
            "description": "Quality of the GPS signal"
          },
          "current_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Lat",
            "description": "Current latitude"
          },
          "current_lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Lng",
            "description": "Current longitude"
          },
          "heading_degrees": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 360.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading Degrees",
            "description": "Robot heading in degrees (0-360)"
          },
          "battery_percent": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Battery Percent",
            "description": "Battery percentage (0-100)"
          },
          "points_treated_today": {
            "type": "integer",
            "title": "Points Treated Today",
            "description": "Number of waypoints treated today",
            "default": 0
          },
          "target_waypoint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Waypoint Id",
            "description": "ObjectId string of the target waypoint, if dispatched"
          },
          "mission_state": {
            "type": "string",
            "enum": [
              "idle",
              "navigating",
              "baseline_evaluating",
              "dispensing_flocculant",
              "incubating_15m",
              "mesh_biochar_filtering",
              "post_evaluating",
              "adaptive_stabilization",
              "completed",
              "failsafe_buoyancy",
              "manual_override"
            ],
            "title": "Mission State",
            "description": "The robot's current mission phase",
            "default": "idle"
          }
        },
        "type": "object",
        "required": [
          "operation_mode",
          "battery_percent"
        ],
        "title": "RobotStatusUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WaypointOut": {
        "properties": {
          "robot_id": {
            "type": "string",
            "title": "Robot Id",
            "description": "The ID of the robot this waypoint belongs to"
          },
          "point_number": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Point Number",
            "description": "1-based index of the waypoint (max 6)"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude",
            "description": "Decimal degrees latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude",
            "description": "Decimal degrees longitude"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "Optional label, defaults to 'Point {point_number}'"
          },
          "radius_meters": {
            "type": "number",
            "title": "Radius Meters",
            "description": "Boundary radius around the waypoint center. Defaults to 2.0",
            "default": 2.0
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "MongoDB ObjectId as string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Server-set UTC timestamp"
          },
          "treated": {
            "type": "boolean",
            "title": "Treated",
            "description": "Whether this waypoint has been treated",
            "default": false
          },
          "treated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Treated At",
            "description": "UTC timestamp when treatment completed"
          },
          "before_reading_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Reading Id",
            "description": "ObjectId string of the before treatment sensor reading"
          },
          "after_reading_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Reading Id",
            "description": "ObjectId string of the after treatment sensor reading"
          }
        },
        "type": "object",
        "required": [
          "robot_id",
          "point_number",
          "latitude",
          "longitude",
          "id",
          "created_at"
        ],
        "title": "WaypointOut"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}